xfoil directory (#95)
* Move xfoil_wrapper.py to xfoil/wrapper.py
* Move XFoilFit class into xfoil/constraint_set.py
